+++ lib/predict/widgets/desktop_prediction_card.dart
