+++ lib/predict/widgets/view_button.dart
